added blank CSS file, blank unit test file, moved functions across to a dedicated functions file
